rename to linkedTo
