+++ crates/sargon-uniffi/src/system/sargon_os/entities_controlled_by_factor_source/sargon_os_entities_linked_to_factor_source.rs
@@ -2,14 +2,14 @@ use crate::prelude::*;
 
 #[uniffi::export]
 impl SargonOS {
-    /// Returns the entities controlled by a given `FactorSource`, either on the current `Profile` or a specific one.
-    pub async fn entities_controlled_by_factor_source(
+    /// Returns the entities linked to a given `FactorSource`, either on the current `Profile` or a specific one.
+    pub async fn entities_linked_to_factor_source(
         &self,
         factor_source: FactorSource,
         profile_to_check: ProfileToCheck,
-    ) -> Result<EntitiesControlledByFactorSource> {
+    ) -> Result<EntitiesLinkedToFactorSource> {
         self.wrapped
-            .entities_controlled_by_factor_source(
+            .entities_linked_to_factor_source(
                 factor_source.into_internal(),
                 profile_to_check.into_internal(),
             )
