+++ crates/sargon-uniffi/src/system/sargon_os/entities_controlled_by_factor_source/support/entities_linked_to_factor_source.rs
@@ -1,21 +1,21 @@
 use crate::prelude::*;
-use sargon::EntitiesControlledByFactorSource as InternalEntitiesControlledByFactorSource;
+use sargon::EntitiesLinkedToFactorSource as InternalEntitiesLinkedToFactorSource;
 
 /// This is the result of checking what entities are controlled by a given `FactorSource`.
 #[derive(Clone, PartialEq, InternalConversion, uniffi::Record)]
-pub struct EntitiesControlledByFactorSource {
+pub struct EntitiesLinkedToFactorSource {
     /// The accessibility of the factor source.
     pub accessibility: FactorSourceAccessibility,
 
-    /// The visible accounts controlled by the factor source.
+    /// The visible accounts linked to the factor source.
     pub accounts: Vec<Account>,
 
-    /// The hidden accounts controlled by the factor source.
+    /// The hidden accounts linked to the factor source.
     pub hidden_accounts: Vec<Account>,
 
-    /// The visible personas controlled by the factor source.
+    /// The visible personas linked to the factor source.
     pub personas: Vec<Persona>,
 
-    /// The hidden personas controlled by the factor source.
+    /// The hidden personas linked to the factor source.
     pub hidden_personas: Vec<Persona>,
 }
