+++ crates/sargon-uniffi/src/system/sargon_os/entities_controlled_by_factor_source/support/mod.rs
@@ -1,7 +1,7 @@
 mod accessibility;
-mod entities_controlled_by_factor_source;
+mod entities_linked_to_factor_source;
 mod profile_to_check;
 
 pub use accessibility::*;
-pub use entities_controlled_by_factor_source::*;
+pub use entities_linked_to_factor_source::*;
 pub use profile_to_check::*;
