+++ crates/sargon-uniffi/src/system/sargon_os/entities_controlled_by_factor_source/support/profile_to_check.rs
@@ -1,7 +1,7 @@
 use crate::prelude::*;
 use sargon::ProfileToCheck as InternalProfileToCheck;
 
-/// The Profile to which we want to check the entities controlled by a factor source.
+/// The Profile to which we want to check the entities linked to a factor source.
 #[derive(Clone, PartialEq, InternalConversion, uniffi::Enum)]
 #[allow(clippy::large_enum_variant)]
 pub enum ProfileToCheck {
