+++ crates/sargon/src/profile/mfa/secured_entity_control/secured_entity_control.rs
@@ -66,8 +66,8 @@ impl SecuredEntityControl {
 }
 
 impl SecuredEntityControl {
-    /// Returns whether the entity is controlled by the given factor source.
-    pub fn is_controlled_by_factor_source(
+    /// Returns whether the entity is linked to the given factor source.
+    pub fn is_linked_to_factor_source(
         &self,
         factor_source: FactorSource,
     ) -> bool {
