+++ crates/sargon/src/profile/v100/entity_security_state/entity_security_state.rs
@@ -86,17 +86,17 @@ impl HasSampleValues for EntitySecurityState {
 }
 
 impl EntitySecurityState {
-    /// Returns whether the entity is controlled by the given factor source.
-    pub fn is_controlled_by_factor_source(
+    /// Returns whether the entity is linked to the given factor source.
+    pub fn is_linked_to_factor_source(
         &self,
         factor_source: FactorSource,
     ) -> bool {
         match self {
             EntitySecurityState::Unsecured { value } => {
-                value.is_controlled_by_factor_source(factor_source)
+                value.is_linked_to_factor_source(factor_source)
             }
             EntitySecurityState::Securified { value } => {
-                value.is_controlled_by_factor_source(factor_source)
+                value.is_linked_to_factor_source(factor_source)
             }
         }
     }
