--- crates/sargon/src/system/sargon_os/entities_controlled_by_factor_source/mod.rs
@@ -1,5 +1,0 @@
-mod sargon_os_entities_controlled_by_factor_source;
-mod support;
-
-pub use sargon_os_entities_controlled_by_factor_source::*;
-pub use support::*;
