+++ crates/sargon/src/system/sargon_os/entities_linked_to_factor_source/support/accessibility/device.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 /// A struct representing the accessibility of a device factor source.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct DeviceFactorSourceAccessibility {
-    /// The factor source that controls the entities.
+    /// The factor source that is linked to the entities.
     pub factor_source: DeviceFactorSource,
 
     /// Whether the mnemonic of the factor source is present in keychain.
