+++ crates/sargon/src/system/sargon_os/entities_linked_to_factor_source/support/profile_to_check.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 
-/// The Profile to which we want to check the entities controlled by a factor source.
+/// The Profile to which we want to check the entities linked to a factor source.
 #[derive(Clone, Debug, PartialEq)]
 #[allow(clippy::large_enum_variant)]
 pub enum ProfileToCheck {
