+++ crates/sargon/src/system/sargon_os/mod.rs
@@ -1,5 +1,5 @@
 mod delete_account;
-mod entities_controlled_by_factor_source;
+mod entities_linked_to_factor_source;
 mod pre_authorization;
 mod profile_state_holder;
 mod sargon_os;
@@ -15,7 +15,7 @@ mod sargon_os_sync_accounts;
 mod transactions;
 
 pub use delete_account::*;
-pub use entities_controlled_by_factor_source::*;
+pub use entities_linked_to_factor_source::*;
 pub use pre_authorization::*;
 pub use profile_state_holder::*;
 pub use sargon_os::*;
